+++ sdk/typescript/src/types/faucet.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { TransactionDigest, ObjectId } from './shared';
+import { TransactionDigest, ObjectId } from './common';
 
 export type FaucetCoinInfo = {
   amount: number;
