+++ sdk/typescript/src/types/framework.ts
@@ -11,7 +11,7 @@ import {
   getMoveObjectType,
   getObjectId,
 } from './objects';
-import { normalizeSuiObjectId } from './common';
+import { normalizeSuiObjectId, ObjectId, SuiAddress } from './common';
 
 import { getOption, Option } from './option';
 import { StructTag } from './sui-bcs';
@@ -19,7 +19,6 @@ import { isSuiMoveObject } from './index.guard';
 import { UnserializedSignableTransaction } from '../signers/txn-data-serializers/txn-data-serializer';
 import { SignerWithProvider } from '../signers/signer-with-provider';
 import { Infer, is, literal, number, object, string, union } from 'superstruct';
-import { ObjectId, SuiAddress } from './shared';
 
 export const SUI_FRAMEWORK_ADDRESS = '0x2';
 export const MOVE_STDLIB_ADDRESS = '0x1';
