+++ sdk/typescript/src/types/index.ts
@@ -9,4 +9,3 @@ export * from './framework';
 export * from './sui-bcs';
 export * from './version';
 export * from './faucet';
-export * from './shared';
