+++ sdk/typescript/src/types/normalized.ts
@@ -1,0 +1,2 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
